LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/drawing-rectangles-to-the-canvas/2d.strokeRect.globalalpha.worker-expected.txt


PASS strokeRect is affected by globalAlpha
